Rename jarvis.py to fuckyalife.py
This uses a couple python modules to listen for the phrase bing bong and then play audio that says "FUCK YA LIFE...BING BONG". it only works when you have the sound file in the same directory as this program. you can find the video on youtube, download it, and then covert it to mp3.